backend/requirements.txt
pydantic
pydantic_settings
sqlalchemy
pandas
numpy
fastapi
requests
uvicorn
psycopg2-binary==2.9.9
pydantic[email]
passlib[bcrypt]
python-jose
